use `simpleBuild`
